+++ Construct/Elements/WordInfo.cs
@@ -60,9 +60,8 @@ namespace PLGL.Construct.Elements
         public enum CaseType { Lowercase, Capitalize, Uppercase, RandomCase }
         public CaseType Case { get; set; } = CaseType.Lowercase;
 
-        public override string ToString()
-        {
-            return WordActual;
-        }
+        public List<LetterInfo> GeneratedLetters { get; set; } = new List<LetterInfo>();
+
+        public override string ToString() { return WordActual; }
     }
 }